+++ node-knx/node-destination.js
@@ -1,18 +1,26 @@
+const knx = require('knx')
+
 module.exports = function (RED) {
 
   function DestinationNode (config) {
     RED.nodes.createNode(this, config)
 
     const node = this
-    const gateway = RED.nodes.getNode(config.gateway)
+    const gateway = config.gateway ? RED.nodes.getNode(config.gateway).gateway : null
     const destination = config.destination
     const dpt = config.dpt
 
     if (gateway && destination) {
-      // Listen to inputs
+      // Connect to the gateway
+      gateway.connect()
+
+      // Create a new data point
+      node.dp = new knx.Datapoint({ga: destination, dpt: dpt, autoread: true}, gateway.connection)
+
+      // Listen to inputs and write values
       node.on('input', function (msg) {
-        if (msg.hasOwnProperty('payload')) {
-          gateway.publish(destination, msg.payload, dpt)
+        if (node.dp && msg.hasOwnProperty('payload')) {
+          node.dp.write(msg.payload.value || msg.payload)
         }
       }.bind(node))
     }